gitlab-ci: add Mono 5.10 stable to new frontend-build phase
This solution includes .NET Standard projects and other
modern things that cannot be compiled in the stock version
of Mono that Ubuntu provides, so we add Xamarin's packages
in this pipeline to at least make sure they pass the build.

It's different from gwallet.sln because it doesn't include
the non-Linux compatible frontends.